requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
--index-url http://artifactory-singlep.p001.alm.brand.dk/artifactory/api/pypi/pypi-virtual/simple
--trusted-host artifactory-singlep.p001.alm.brand.dk

attrs==19.2.0
cycler==0.10.0            # via matplotlib
gitdb2==2.0.5             # via gitpython
gitpython==3.0.2
joblib==0.13.2
kiwisolver==1.1.0         # via matplotlib
matplotlib==3.1.1
numpy==1.17.2             # via matplotlib, pandas, pyarrow, scikit-learn, scipy
pandas==0.25.1
pyarrow==0.14.1
pyparsing==2.4.2          # via matplotlib
python-dateutil==2.8.0    # via matplotlib, pandas
pytz==2019.2              # via pandas
pyyaml==5.1.2
scikit-learn==0.21.3
scipy==1.3.1              # via scikit-learn
six==1.12.0               # via cycler, pyarrow, python-dateutil
smmap2==2.0.5             # via gitdb2
sqlalchemy==1.3.8

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.4.0        # via kiwisolver
